more cautious mode

+++ R/statistics.R
@@ -19,6 +19,7 @@
 #' @param verbose Logical, print verbose messages.
 #' @param cores number of cores used when computing profiles for several
 #' parameters.
+#' @param cautiousMode Logical, write every step to disk and don't delete intermediate results
 #' @param ... Arguments going to obj()
 #' @details Computation of the profile likelihood is based on the method of Lagrangian multipliers
 #' and Euler integration of the corresponding differential equation of the profile likelihood paths.
@@ -55,6 +56,7 @@ profile <- function(obj, pars, whichPar, alpha = 0.05,
                     optControl  = NULL,
                     verbose = FALSE,
                     cores = 1,
+                    cautiousMode = FALSE,
                     ...) {
   
   # Guarantee that pars is named numeric without deriv attribute
@@ -87,9 +89,11 @@ profile <- function(obj, pars, whichPar, alpha = 0.05,
   if (!is.null(optControl )) oControl[match(names(optControl), names(oControl ))] <- optControl
   
   
-  
-  
-  
+  # Create interRes folder for cautiousMode
+  if (cautiousMode){
+    profiles_interResFolder <- "profiles-interRes"
+    dir.create(profiles_interResFolder,showWarnings = FALSE)
+  }
   
   
   # Start cluster if on windows
@@ -410,6 +414,9 @@ profile <- function(obj, pars, whichPar, alpha = 0.05,
                                              out.attributes, 
                                              y))
                               
+                              if(cautiousMode) {
+                              }
+                              
                               value <- lagrange.out[[sControl$stop]]
                               if (value > threshold | constraint.out$value > limits[2]) break
                               
@@ -670,7 +677,7 @@ vcov <- function(fit, parupper = NULL, parlower = NULL) {
   subvcov__ <- try(solve(0.5*subhessian__), silent = TRUE)
   if (inherits(subvcov__, "try-error")) subvcov__ <- MASS::ginv(subhessian__)
   vcov__[!is_fixed__, !is_fixed__] <- subvcov__
-
+  
   # This part should not be necessary due to regularization usually done
   # Perform identifiability check based on
   # subvcov__ <- vcov__[!is_fixed__, !is_fixed__, drop = FALSE]
@@ -724,6 +731,7 @@ vcov <- function(fit, parupper = NULL, parlower = NULL) {
 #'   are handed to the objective function objfun(). The log file starts with a 
 #'   table telling which parameter was assigend to which function.
 #' @param output logical. If true, writes output to the disc.
+#' @param cautiousMode Logical, write every fit to disk in deparsed form (avoids the RDA incompatibility trap) and don't delete intermediate results
 #'   
 #' @details By running multiple fits starting at randomly chosen inital 
 #'   parameters, the chisquare landscape can be explored using a deterministic 
@@ -764,7 +772,7 @@ vcov <- function(fit, parupper = NULL, parlower = NULL) {
 #' @export
 #' @import parallel
 mstrust <- function(objfun, center, studyname, rinit = .1, rmax = 10, fits = 20, cores = 1, optmethod = "trust",
-                    samplefun = "rnorm", resultPath = ".", stats = FALSE, output = FALSE,
+                    samplefun = "rnorm", resultPath = ".", stats = FALSE, output = FALSE, cautiousMode = FALSE,
                     ...) {
   
   narrowing <- NULL
@@ -958,7 +966,7 @@ mstrust <- function(objfun, center, studyname, rinit = .1, rmax = 10, fits = 20,
     # Write current fit to disk
     if (output) {
       saveRDS(fit, file = file.path(interResultFolder, paste0("fit-", i, ".Rda")))
-      dput(fit, file = file.path(interResultFolder, paste0("fit-", i, ".R")))
+      if (cautiousMode) dput(fit, file = file.path(interResultFolder, paste0("fit-", i, ".R")))
       # Reporting
       # With concurent jobs and everyone reporting, this is a classic race
       # condition. Assembling the message beforhand lowers the risk of interleaved
@@ -1030,7 +1038,7 @@ mstrust <- function(objfun, center, studyname, rinit = .1, rmax = 10, fits = 20,
   if (output) saveRDS(m_parlist, file = fileParList)
   
   # Remove temporary files
-  unlink(interResultFolder, recursive = TRUE)
+  if (!cautiousMode) unlink(interResultFolder, recursive = TRUE)
   
   
   # Show summary
